switch from oneof to any for transaction payload fields
Because oneof fields in Golang are defined in terms of an interface instead of a
concrete type definition, CBOR has trouble decoding these fields and finding
their correct concrete type. Any fields are a byte array wrapped in a concrete
type along with metadata, so CBOR should be able to decode them better on the go
side.